requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
psycopg[binary]==3.3.2
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.3.0
python-multipart==0.0.20
alembic==1.14.0
email-validator==2.2.0

# Testing
pytest==8.3.4
pytest-cov==6.0.0
httpx==0.28.1
